# 🚀 Enhanced API Functionality and Performance Improvements